--- takoyaki_bin/src/middlewares/mod.rs
@@ -1,3 +1,0 @@
-mod auth;
-
-pub use auth::*;
