--- takoyaki_bin/src/routes/mod.rs
@@ -1,6 +1,0 @@
-mod deploy;
-mod poll;
-
-pub use deploy::*;
-pub use poll::*;
-
