--- takoyaki_bin/src/utils/mod.rs
@@ -1,8 +1,0 @@
-mod deploy;
-mod executor;
-mod directories;
-
-pub use deploy::*;
-pub use executor::*;
-pub use directories::*;
-
