+++ takoyaki_plugin_deployer/.gitignore
@@ -1,1 +1,1 @@
-target/
+target
