+++ takoyaki_plugin_deployer/src/routes/mod.rs
@@ -1,3 +1,6 @@
 mod deploy;
+mod poll;
 
 pub use deploy::*;
+pub use poll::*;
+
