+++ takoyaki_plugin_deployer/src/routes/poll.rs
